PortfolioViewService on frontend implemented
