+++ frontend/src/main/java/de/iav/frontend/service/UserService.java
@@ -2,7 +2,6 @@ package de.iav.frontend.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.iav.frontend.model.TransactionWithoutUser;
 import de.iav.frontend.model.User;
 import java.net.URI;
 import java.net.http.HttpClient;
@@ -10,79 +9,51 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.util.List;
 public class UserService {
+
     private static UserService instance;
     private final HttpClient httpClient;
     private final ObjectMapper objectMapper;
+
     private UserService() {
         this.httpClient = HttpClient.newHttpClient();
         this.objectMapper = new ObjectMapper();
     }
+
     public static synchronized UserService getInstance() {
         if (instance == null) {
             instance = new UserService();
         }
         return instance;
     }
+
     public List<User> getAllUsers() {
         HttpRequest request = HttpRequest.newBuilder()
                 .GET()
                 .uri(URI.create("http://localhost:8080/api/financeapp/users"))
                 .build();
+
         return httpClient.sendAsync(request, HttpResponse.BodyHandlers.ofString())
                 .thenApply(HttpResponse::body)
                 .thenApply(this::mapToUserList) // .thenApply(responseBody -> mapToStudent(responseBody))
                 .join();
     }
+
     private List<User> mapToUserList(String responseBody) {
         try {
             return objectMapper.readValue(responseBody, new TypeReference<>() {
             });
+            //return new ArrayList<Student>();
         } catch (JsonProcessingException e) {
             throw new RuntimeException("Failed to map users List", e);
         }
+
     }
+
     private User mapToUser(String json) {
         try {
             return objectMapper.readValue(json, User.class);
         } catch (JsonProcessingException e) {
             throw new RuntimeException("Failed to map stock", e);
-        }
-    }
-
-    public List<TransactionWithoutUser> getPortfolioForUser(String userId) {
-        HttpRequest request = HttpRequest.newBuilder()
-                .GET()
-                .uri(URI.create("http://localhost:8080/api/financeapp/users/" + userId + "/portfolio"))
-                .build();
-        return httpClient.sendAsync(request, HttpResponse.BodyHandlers.ofString())
-                .thenApply(HttpResponse::body)
-                .thenApply(this::mapToPortfolioList)
-                .join();
-    }
-    private List<TransactionWithoutUser> mapToPortfolioList(String responseBody) {
-        try {
-            return objectMapper.readValue(responseBody, new TypeReference<>() {});
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Failed to map portfolio list", e);
-        }
-    }
-
-    public Double getPortfolioValue(String userId) {
-        HttpRequest request = HttpRequest.newBuilder()
-                .GET()
-                .uri(URI.create("http://localhost:8080/api/financeapp/users/" + userId + "/portfoliovalue"))
-                .build();
-        return httpClient.sendAsync(request, HttpResponse.BodyHandlers.ofString())
-                .thenApply(HttpResponse::body)
-                .thenApply(this::mapToPortfolioValue)
-                .join();
-    }
-
-    private Double mapToPortfolioValue(String responseBody) {
-        try {
-            return objectMapper.readValue(responseBody, new TypeReference<>() {});
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("Failed to map portfolio value", e);
         }
     }
 }
